+++ MatchEntities/Models/PlayerPosition.cs
@@ -10,14 +10,14 @@ namespace MatchEntities
         public int Time { get; set; }
         public int Tick { get; set; }
         public long PlayerId { get; set; }
-        public double PlayerPosX { get; set; }
-        public double PlayerPosY { get; set; }
-        public double PlayerPosZ { get; set; }
-        public double PlayerViewX { get; set; }
-        public double PlayerViewY { get; set; }
-        public double PlayerVeloX { get; set; }
-        public double PlayerVeloY { get; set; }
-        public double PlayerVeloZ { get; set; }
+        public float PlayerPosX { get; set; }
+        public float PlayerPosY { get; set; }
+        public float PlayerPosZ { get; set; }
+        public float PlayerViewX { get; set; }
+        public float PlayerViewY { get; set; }
+        public float PlayerVeloX { get; set; }
+        public float PlayerVeloY { get; set; }
+        public float PlayerVeloZ { get; set; }
         public short Weapon { get; set; }
 
         public MatchStats MatchStats { get; set; }
